# replace rng with deterministic indexer