Closes https://github.com/yoshuawuyts/futures-concurrency/issues/85. Additionally it also significantly simplifies our iteration logic by baking it into the `Indexer` struct directly:

```rust
// before
let len = this.streams.len();
let r = this.rng.generate(len as u32) as usize;
for index in (0..len).map(|n| (r + n).wrapping_rem(len)) { ... }

// after
for index in this.indexer.iter() { ... }
```

## Performance

The numbers overall seem pretty comparable, but they spiked a bit on non-allocating APIs which are otherwise fairly simple. `impl Race for array` showed one of the bigger changes:
```text
❯ critcmp main patch
group             main                    patch
-----             ----                    -----
array::race 10    1.22   652.2±24.21ns    1.00   535.9±31.39ns
array::race 100   1.52      4.8±0.27µs    1.00      3.2±0.02µs
array::race 1000  1.07     28.4±0.56µs    1.00     26.6±0.58µs
```